DELIA-53086 : Improve wifi sec/encryp tag parser and testing
Reason for change: We should be able to connect if FT is enabled on AP.
Also with this change adding desktop build and google test support
so that we can experiment and test scenarios like this more easily.
Also cleaned up many compile warnings.
Test Procedure: Ensure ticket is resolved and no new regressions
Risks: Low
Signed-off-by: mrollins <mark_rollins@cable.comcast.com>
Change-Id: I4fa220fef1f7a1bf68d0787eff63f38e1b3a55e7